build(mise): Add entries for hugo-extended and golang
https://mise.jdx.dev/ is an "asdf" alternative I use that also reads
from the .tool-versions file.

This change:

- Adds an entry for `golang`. Mise isn't written in Go, so you can't
assume Go is already installed on the user's system.

- Adds an entry for `hugo-extended`. Because we don't deserve nice
things the asdf folks appear to have decided that `extended` is part
of the version component, while the mise folks decided it's part of
the package name.

With this change someone who uses Mise can `cd` to the `website`
directory and `mise` will ensure the correct tools are installed so
that `make dev` works immediately.